==> 02_exploratory_analysis.Rmd <==
```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

source("helper_functions/exploratory_analysis.R")
```

```{r}
h1b_df <- read.csv("clean_data/h1b_FY2021_FY2024.csv")
```

## Context

This analysis estimates lottery win rates and multiple registration rates by country, company, and age. While differences in win rates by these segments might look like evidence of discrimination, this inference is not necessarily accurate unless one can convincingly rule out confounding by degree level. 

## How does the win rate vary by country? 

```{r}
country_code_mapping <- list(CAN = "Canada", 
                             CHN = "China", 
                             GBR = "United Kingdom", 
                             IND = "India", 
                             KOR = "South Korea", 
                             MEX = "Mexico", 
                             NPL = "Nepal", 
                             PAK = "Pakistan", 
                             PHL = "Philippines", 
                             TWN = "Taiwan", 
                             SGP = "Singapore", 
                             CHL = "Chile")

bump_plot(h1b_df, "country_code", "Countries", "won_lottery", "Lottery Win Rates", 10, country_code_mapping, "Certain Countries Look Like They Have An Advantage In The H-1B Visa Lottery")
```

## How does the multiple registration rate vary by country? 

```{r}
line_plot_categorical(h1b_df, "country_code", "Countries", "had_multiple_regs", "Multiple Registration Rates", 10, country_code_mapping, "India Has Higher Multiple Registration Rates In The H-1B Visa Lottery")
```

## How does the win rate vary by company? 

```{r}
company_mapping <- list(`Amazon.com Services LLC` = "Amazon Services", 
                             `Google LLC` = "Google", 
                             `Microsoft Corporation` = "Microsoft", 
                             `Capgemini America, Inc.` = "Capgemini", 
                             `Tata Consultancy Services Limited` = "Tata", 
                             `HCL America Inc` = "HCL", 
                             `Cognizant Technology Solutions US Corp` = "Cognizant", 
                             `IBM Corporation` = "IBM", 
                             `IBM Corp` = "IBM", 
                             `Infosys Limited` = "Infosys", 
                             `Accenture LLP` = "Accenture", 
                             `Wipro Limited` = "Wipro", 
                             `WIPRO LIMITED` = "Wipro", 
                             `Mindtree Limited` = "LTIMindtree", 
                             `MindTree Limited` = "LTIMindtree", 
                             `LTIMINDTREE LIMITED` = "LTIMindtree", 
                             `Meta Platforms, Inc.` = "Meta", 
                             `Facebook, Inc.` = "Meta", 
                             `Tech Mahindra Americas Inc` = "Tech Mahindra", 
                             `Tech Mahindra Americas Inc.` = "Tech Mahindra")


bump_plot(h1b_df, "employer_name", "Companies", "won_lottery", "Lottery Win Rates", 10, company_mapping, 
          "Certain Companies Look Like They Have An Advantage In The H-1B Visa Lottery")
```

## How does the multiple registration rate vary by company? 

```{r}
line_plot_categorical(h1b_df, "employer_name", "Companies", "had_multiple_regs", "Multiple Registration Rates", 10, company_mapping, "Multiple Registration Rates By Company Fluctuate Considerably Across Years")
```

## How does the win rate vary by age? 

```{r}
line_plot_continuous(h1b_df, "age_in_years", "Age", "won_lottery", "Win", 21, 50, "25-30 Year Olds Look Like They Have An Advantage in the H-1B Visa Lottery")
```

## How does the multiple registration rate vary by age? 

```{r}
line_plot_continuous(h1b_df, "age_in_years", "Age", "had_multiple_regs", "Multiple Registration", 21, 50, "Multiple Registration Rates Vary By Age")
```

## How has the volume of multiple registrants changed over time? 

```{r}
reg_df <- read.csv("clean_data/h1b_volume_FY2021_FY2025.csv")

ggplot(reg_df, aes(x = lottery_year, y = num_registrations, group = registration_type, fill = registration_type))+
  geom_bar(stat = "identity")+
  geom_vline(xintercept = 2024.5, linetype = "dashed")+
  labs(x = "Lottery Year (FY)", y = "# of Lottery Registrations", title = "USCIS Policy Change Substantially Reduced Multiple Registrations", caption = "Visualization: Vinay Tummarakota, @unboxpolitics\nSource: USCIS")+
  scale_y_continuous(labels = label_number(scale_cut = cut_short_scale()))+
  scale_fill_discrete(name = "Registration Type")+
  theme_minimal()
```

## How often do lottery winners have missing data on degree level? 

```{r}
missing_edu_df <- h1b_df %>%
  filter(won_lottery == TRUE) %>%
  group_by(lottery_year) %>%
  summarise(missing_data_rate = sum(degree_level_is_missing)/n())

ggplot(missing_edu_df, aes(x = lottery_year, y = missing_data_rate))+
  geom_bar(stat = "identity")+
  labs(x = "Lottery Year (FY)", y = "Missing Data Rate", title = "20-60% of Lottery Winners Don't Have Data on Degree Level", caption = "Visualization: Vinay Tummarakota, @unboxpolitics\nSource: USCIS (Obtained by Bloomberg)")+
  scale_y_continuous(labels = percent_format())+
  theme_minimal()
```

## Is missing degree level data disproportionately concentrated in large IT staffing firms? 

```{r}
h1b_winners_df <- h1b_df %>% 
  filter(won_lottery == TRUE)

line_plot_categorical(h1b_winners_df, "employer_name", "Companies", "degree_level_is_missing", "Missing Data Rate", 10, company_mapping, "Missing Data Rates By Company Fluctuate Considerably Across Years")
```

## Were any H-1B visas given to janitors or cashiers?  

```{r}
h1b_df %>%
  filter(petition_approved == TRUE, dot_code %in% c("211", "382") | grepl("janitor|cashier|custodia", tolower(job_title))) %>%
  arrange(annual_salary) %>%
  select(lottery_year, employer_name, job_title, annual_salary) %>%
  top_n(n = 5)
```

## What fraction of H-1B visa acceptances were for salaries >=60k/year? 

```{r}
salary_df <- h1b_df %>%
  filter(petition_approved == TRUE) %>%
  mutate(annual_salary_above_60k = annual_salary >= 60000) %>%
  group_by(lottery_year) %>%
  summarise(share_annual_salary_above_60k = sum(annual_salary_above_60k, na.rm = TRUE) / n())

ggplot(salary_df, aes(x = lottery_year, y = share_annual_salary_above_60k))+
  geom_bar(stat = "identity")+
  geom_text(aes(label = glue("{100*round(share_annual_salary_above_60k, 3)}%"), vjust = -0.2))+
  labs(x = "Lottery Year (FY)", y = "% Listed Salary At Least $60k/Year", title = "Vast majority of H-1B visa acceptances have listed salary\nat least $60k/year", subtitle = "Sample is limited to H-1B visa acceptances", caption = "Visualization: Vinay Tummarakota, @unboxpolitics\nSource: USCIS (Obtained by Bloomberg)")+
  scale_y_continuous(labels = percent_format(), limits = c(0, 1.05))+
  theme_minimal()
```
